refactoring to get rid of chat-video
