+++ app/components/shared/avatar.js
@@ -19,7 +19,7 @@ import { User } from '../../lib/icebear';
 import { tx } from '../utils/translator';
 import preferenceStore from '../settings/preference-store';
 import { popupSetupVideo } from '../shared/popups';
-import ChatVideo from '../messaging/chat-video';
+import VideoIcon from '../layout/video-icon';
 
 
 const itemStyle = {
@@ -188,7 +188,7 @@ export default class Avatar extends SafeComponent {
     }
 
     popupCall(video) {
-        return popupSetupVideo(ChatVideo.storeLink, video);
+        return popupSetupVideo(VideoIcon.storeLink, video);
     }
 
     get systemMessage() {
